pyproject.toml
[project]
name = "stein-sense"
version = "0.0.0"
description = "Default template for the Bayesian sensitivity analysis package"
authors = [
    { name = "Arina Odnoblyudova", email = "amodnoblyudova@gmail.com" },
]
dependencies = [
    "numpy~=1.24",
    "scipy~=1.10",
    "pydantic~=2.7",
    "matplotlib~=3.7",
    "scikit-image~=0.21",
    "scikit-learn~=1.3",
    "pre-commit~=4.2",
    "abstract>=2025.2.17.0",
    "posteriordb>=0.2.0",
    "hydra-core>=1.3.2",
    "pymc>=5.23.0",
    "seaborn>=0.13.2",
]
requires-python = "~=3.12.0"
readme = "README.md"
license = { text = "Proprietary" }